+++ src/Tests/Bastion.fs
@@ -8,6 +8,7 @@ open Farmer.Builders
 open Microsoft.Azure.Management.Network
 open Microsoft.Azure.Management.Network.Models
 open Microsoft.Rest
+open Newtonsoft.Json.Linq
 
 /// Client instance needed to get the serializer settings.
 let client =
@@ -56,5 +57,47 @@ let tests =
                     resources.[1].IpConfigurations.[0].Subnet.Id
                     "[resourceId('Microsoft.Network/virtualNetworks/subnets', 'private-network', 'AzureBastionSubnet')]"
                     "Subnet name must be 'AzureBastionSubnet'"
+            }
+            test "Advanced settings upgrade to standard sku" {
+                let deployment =
+                    arm {
+                        location Location.EastUS
+
+                        add_resources
+                            [
+                                bastion {
+                                    name "my-bastion"
+                                    link_to_vnet "my-vnet"
+                                    enable_shareable_link true
+                                    enable_tunneling true
+                                }
+                            ]
+                    }
+
+                let jobj = deployment.Template |> Writer.toJson |> JToken.Parse
+                let bastion = jobj.SelectToken "resources[?(@.name == 'my-bastion')]"
+                Expect.isNotNull bastion "Bastion missing from template"
+
+                Expect.equal
+                    (bastion.SelectToken "sku.name" |> string)
+                    "Standard"
+                    "Bastion SKU should be upgraded to Standard"
+
+                Expect.isTrue
+                    ((bastion.SelectToken "properties.enableShareableLink").ToObject<bool>())
+                    "Bastion should have shareable link enabled"
+
+                Expect.isTrue
+                    ((bastion.SelectToken "properties.enableTunneling").ToObject<bool>())
+                    "Bastion should have tunneling enabled"
+
+                Expect.isTrue
+                    ((bastion.SelectToken "properties.enableTunneling").ToObject<bool>())
+                    "Bastion should have tunneling enabled"
+
+                Expect.equal
+                    ((bastion.SelectToken "properties.scaleUnits").ToObject<int>())
+                    2
+                    "Bastion should specify 2 scale units by default"
             }
         ]